refactor(web): split lib/api.ts by concern, move jsonApi to src/helpers
api.ts mixed a token store, an ApiError class, an HTTP client, and auth
actions in one file. Split into accessToken.ts, ApiError.ts, httpClient.ts,
and auth.ts — one concern each. jsonApi.ts moves from src/lib to
src/helpers: it's pure envelope-parsing with no state, not a lib module.
curriculum.ts's type/interface declarations now all precede its functions.
