+++ apps/web/src/lib/routeGuards.ts
@@ -1,5 +1,5 @@
 import { redirect } from "@tanstack/react-router";
-import { getAccessToken } from "./api";
+import { getAccessToken } from "./accessToken";
 
 // Shared `beforeLoad` guard for every authenticated route.
 export function requireAuth() {
